Initial implementation of the FreeDrag plugin
This plugin is intended to allow dragging visually selected blocks of
text in any of four directions, in a manner that has a predictable
impact on the characters around it. In particular, the line of text in
the direction of the move gets displaced and nothing else moves.
This initial commit only contains the implementation for moving up.